--- src/SchemaBuilder/ClassDecorator/AnnotationClassSchemaDecorator.php
@@ -1,81 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace ScrumWorks\OpenApiSchema\SchemaBuilder\ClassDecorator;
-
-use Doctrine\Common\Annotations\Reader;
-use ReflectionClass;
-use ReflectionProperty;
-use ScrumWorks\OpenApiSchema\Annotation as OA;
-use ScrumWorks\OpenApiSchema\Exception\LogicException;
-use ScrumWorks\OpenApiSchema\SchemaBuilder\ClassSchemaDecoratorInterface;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\AbstractSchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\ObjectSchemaBuilder;
-
-class AnnotationClassSchemaDecorator implements ClassSchemaDecoratorInterface
-{
-    use ClassSchemaDecoratorDefaultTrait;
-
-    private Reader $annotationReader;
-
-    public function __construct(Reader $annotationReader)
-    {
-        $this->annotationReader = $annotationReader;
-    }
-
-    public function decorateObjectSchemaBuilder(
-        ObjectSchemaBuilder $builder,
-        ReflectionClass $classReflection
-    ): AbstractSchemaBuilder {
-        $objectDefaultValues = $classReflection->getDefaultProperties();
-        $requiredProperties = [];
-        foreach ($classReflection->getProperties() as $propertyReflection) {
-            if ($this->isPropertyRequired($propertyReflection, $objectDefaultValues)) {
-                $requiredProperties[] = $propertyReflection->getName();
-            }
-        }
-        return $builder->withRequiredProperties($requiredProperties);
-    }
-
-    private function isPropertyRequired(ReflectionProperty $propertyReflection, array $objectDefaultValues): bool
-    {
-        $annotations = $this->getPropertyAnnotations($propertyReflection);
-        /** @var ?OA\Property $annotation */
-        $annotation = $this->findAnnotation($annotations, OA\Property::class, false);
-        if ($annotation && $annotation->required !== null) {
-            return $annotation->required;
-        }
-        return ! \array_key_exists($propertyReflection->getName(), $objectDefaultValues);
-    }
-
-    private function getPropertyAnnotations(ReflectionProperty $propertyReflection): array
-    {
-        return $this->annotationReader->getPropertyAnnotations($propertyReflection);
-    }
-
-    private function findAnnotation(
-        array $annotations,
-        string $annotationClass,
-        bool $exceptionOnAnotherValueInterface = true
-    ): ?object {
-        $found = null;
-        foreach ($annotations as $annotation) {
-            if (\get_class($annotation) === $annotationClass) {
-                // micro-optimalization
-                if (! $exceptionOnAnotherValueInterface) {
-                    return $annotation;
-                }
-
-                $found = $annotation;
-            } elseif (
-                $exceptionOnAnotherValueInterface
-                && \is_subclass_of($annotation, OA\ValueInterface::class)
-            ) {
-                throw new LogicException(\sprintf("Unexpected annotation '%s'", \get_class($annotation)));
-            }
-        }
-
-        return $found;
-    }
-}
