--- src/SchemaBuilder/ClassDecorator/ClassSchemaDecoratorDefaultTrait.php
@@ -1,90 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace ScrumWorks\OpenApiSchema\SchemaBuilder\ClassDecorator;
-
-use ReflectionClass;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\AbstractSchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\ArraySchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\BooleanSchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\EnumSchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\FloatSchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\HashmapSchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\IntegerSchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\MixedSchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\ObjectSchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\StringSchemaBuilder;
-
-trait ClassSchemaDecoratorDefaultTrait
-{
-    public function decorateValueSchemaBuilder(
-        AbstractSchemaBuilder $builder,
-        ReflectionClass $classReflection
-    ): AbstractSchemaBuilder {
-        return $builder;
-    }
-
-    public function decorateMixedSchemaBuilder(
-        MixedSchemaBuilder $builder,
-        ReflectionClass $classReflection
-    ): AbstractSchemaBuilder {
-        return $builder;
-    }
-
-    public function decorateIntegerSchemaBuilder(
-        IntegerSchemaBuilder $builder,
-        ReflectionClass $classReflection
-    ): AbstractSchemaBuilder {
-        return $builder;
-    }
-
-    public function decorateFloatSchemaBuilder(
-        FloatSchemaBuilder $builder,
-        ReflectionClass $classReflection
-    ): AbstractSchemaBuilder {
-        return $builder;
-    }
-
-    public function decorateBooleanSchemaBuilder(
-        BooleanSchemaBuilder $builder,
-        ReflectionClass $classReflection
-    ): AbstractSchemaBuilder {
-        return $builder;
-    }
-
-    public function decorateStringSchemaBuilder(
-        StringSchemaBuilder $builder,
-        ReflectionClass $classReflection
-    ): AbstractSchemaBuilder {
-        return $builder;
-    }
-
-    public function decorateEnumSchemaBuilder(
-        EnumSchemaBuilder $builder,
-        ReflectionClass $classReflection
-    ): AbstractSchemaBuilder {
-        return $builder;
-    }
-
-    public function decorateArraySchemaBuilder(
-        ArraySchemaBuilder $builder,
-        ReflectionClass $classReflection
-    ): AbstractSchemaBuilder {
-        return $builder;
-    }
-
-    public function decorateHashmapSchemaBuilder(
-        HashmapSchemaBuilder $builder,
-        ReflectionClass $classReflection
-    ): AbstractSchemaBuilder {
-        return $builder;
-    }
-
-    public function decorateObjectSchemaBuilder(
-        ObjectSchemaBuilder $builder,
-        ReflectionClass $classReflection
-    ): AbstractSchemaBuilder {
-        return $builder;
-    }
-}
