+++ src/SchemaBuilder/Decorator/ClassSchemaDecoratorInterface.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ScrumWorks\OpenApiSchema\SchemaBuilder\Decorator;
+
+use ReflectionClass;
+use ScrumWorks\OpenApiSchema\ValueSchema\Builder\ObjectSchemaBuilder;
+
+interface ClassSchemaDecoratorInterface
+{
+    public function decorateObjectSchemaBuilder(
+        ObjectSchemaBuilder $builder,
+        ReflectionClass $classReflection
+    ): ObjectSchemaBuilder;
+}
