+++ src/SchemaBuilder/Decorator/PropertyDecorator/AnnotationPropertySchemaDecorator.php
@@ -2,41 +2,59 @@
 
 declare(strict_types=1);
 
-namespace ScrumWorks\OpenApiSchema\SchemaBuilder\PropertyDecorator;
+namespace ScrumWorks\OpenApiSchema\SchemaBuilder\Decorator\PropertyDecorator;
 
-use Doctrine\Common\Annotations\Reader;
 use ReflectionProperty;
 use ScrumWorks\OpenApiSchema\Annotation as OA;
-use ScrumWorks\OpenApiSchema\Exception\LogicException;
-use ScrumWorks\OpenApiSchema\SchemaBuilder\PropertySchemaDecoratorInterface;
+use ScrumWorks\OpenApiSchema\SchemaBuilder\Decorator\AbstractAnnotationSchemaDecorator;
+use ScrumWorks\OpenApiSchema\SchemaBuilder\Decorator\PropertySchemaDecoratorInterface;
 use ScrumWorks\OpenApiSchema\ValueSchema\Builder\AbstractSchemaBuilder;
 use ScrumWorks\OpenApiSchema\ValueSchema\Builder\ArraySchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\BooleanSchemaBuilder;
 use ScrumWorks\OpenApiSchema\ValueSchema\Builder\EnumSchemaBuilder;
 use ScrumWorks\OpenApiSchema\ValueSchema\Builder\FloatSchemaBuilder;
 use ScrumWorks\OpenApiSchema\ValueSchema\Builder\HashmapSchemaBuilder;
 use ScrumWorks\OpenApiSchema\ValueSchema\Builder\IntegerSchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\MixedSchemaBuilder;
 use ScrumWorks\OpenApiSchema\ValueSchema\Builder\StringSchemaBuilder;
 
-class AnnotationPropertySchemaDecorator implements PropertySchemaDecoratorInterface
+final class AnnotationPropertySchemaDecorator extends AbstractAnnotationSchemaDecorator implements PropertySchemaDecoratorInterface
 {
-    use PropertySchemaDecoratorDefaultTrait;
-
-    private Reader $annotationReader;
-
-    public function __construct(Reader $annotationReader)
-    {
-        $this->annotationReader = $annotationReader;
-    }
-
-    public function decorateValueSchemaBuilder(
+    public function decoratePropertySchemaBuilder(
         AbstractSchemaBuilder $builder,
         ReflectionProperty $propertyReflection
     ): AbstractSchemaBuilder {
         $annotations = $this->getPropertyAnnotations($propertyReflection);
 
-        if ($annotation = $this->findAnnotation($annotations, OA\Property::class, false)) {
+        return $this->decoratePropertySchemaBuilderFromAnnotations($builder, $annotations);
+    }
+
+    public function decoratePropertySchemaBuilderFromAnnotations(
+        AbstractSchemaBuilder $builder,
+        array $annotations
+    ): AbstractSchemaBuilder {
+        $builder = $this->decorateValueSchemaBuilder($builder, $annotations);
+
+        if ($builder instanceof IntegerSchemaBuilder) {
+            $builder = $this->decorateIntegerSchemaBuilder($builder, $annotations);
+        } elseif ($builder instanceof FloatSchemaBuilder) {
+            $builder = $this->decorateFloatSchemaBuilder($builder, $annotations);
+        } elseif ($builder instanceof StringSchemaBuilder) {
+            $builder = $this->decorateStringSchemaBuilder($builder, $annotations);
+        } elseif ($builder instanceof EnumSchemaBuilder) {
+            $builder = $this->decorateEnumSchemaBuilder($builder, $annotations);
+        } elseif ($builder instanceof ArraySchemaBuilder) {
+            $builder = $this->decorateArraySchemaBuilder($builder, $annotations);
+        } elseif ($builder instanceof HashmapSchemaBuilder) {
+            $builder = $this->decorateHashmapSchemaBuilder($builder, $annotations);
+        }
+
+        return $builder;
+    }
+
+    private function decorateValueSchemaBuilder(
+        AbstractSchemaBuilder $builder,
+        array $annotations
+    ): AbstractSchemaBuilder {
+        if ($annotation = $this->findAnnotation($annotations, OA\Property ::class, false)) {
             /** @var OA\Property $annotation */
             if ($annotation->description !== null) {
                 $builder = $builder->withDescription($annotation->description);
@@ -50,19 +68,10 @@ class AnnotationPropertySchemaDecorator implements PropertySchemaDecoratorInterf
         return $builder;
     }
 
-    public function decorateMixedSchemaBuilder(
-        MixedSchemaBuilder $builder,
-        ReflectionProperty $propertyReflection
-    ): AbstractSchemaBuilder {
-        return $builder;
-    }
-
-    public function decorateIntegerSchemaBuilder(
+    private function decorateIntegerSchemaBuilder(
         IntegerSchemaBuilder $builder,
-        ReflectionProperty $propertyReflection
+        array $annotations
     ): AbstractSchemaBuilder {
-        $annotations = $this->getPropertyAnnotations($propertyReflection);
-
         if ($annotation = $this->findAnnotation($annotations, OA\IntegerValue::class)) {
             /** @var OA\IntegerValue $annotation */
             if ($annotation->minimum !== null) {
@@ -85,12 +94,10 @@ class AnnotationPropertySchemaDecorator implements PropertySchemaDecoratorInterf
         return $builder;
     }
 
-    public function decorateFloatSchemaBuilder(
+    private function decorateFloatSchemaBuilder(
         FloatSchemaBuilder $builder,
-        ReflectionProperty $propertyReflection
+        array $annotations
     ): AbstractSchemaBuilder {
-        $annotations = $this->getPropertyAnnotations($propertyReflection);
-
         if ($annotation = $this->findAnnotation($annotations, OA\FloatValue::class)) {
             /** @var OA\FloatValue $annotation */
             if ($annotation->minimum !== null) {
@@ -113,23 +120,14 @@ class AnnotationPropertySchemaDecorator implements PropertySchemaDecoratorInterf
         return $builder;
     }
 
-    public function decorateBooleanSchemaBuilder(
-        BooleanSchemaBuilder $builder,
-        ReflectionProperty $propertyReflection
-    ): AbstractSchemaBuilder {
-        return $builder;
-    }
-
-    public function decorateStringSchemaBuilder(
+    private function decorateStringSchemaBuilder(
         StringSchemaBuilder $builder,
-        ReflectionProperty $propertyReflection
+        array $annotations
     ): AbstractSchemaBuilder {
-        $annotations = $this->getPropertyAnnotations($propertyReflection);
-
         if ($this->findAnnotation($annotations, OA\EnumValue::class, false)) {
             $builder = (new EnumSchemaBuilder())
                 ->withNullable($builder->isNullable());
-            return $this->decorateEnumSchemaBuilder($builder, $propertyReflection);
+            return $this->decorateEnumSchemaBuilder($builder, $annotations);
         }
 
         if ($annotation = $this->findAnnotation($annotations, OA\StringValue::class)) {
@@ -151,12 +149,10 @@ class AnnotationPropertySchemaDecorator implements PropertySchemaDecoratorInterf
         return $builder;
     }
 
-    public function decorateEnumSchemaBuilder(
+    private function decorateEnumSchemaBuilder(
         EnumSchemaBuilder $builder,
-        ReflectionProperty $propertyReflection
+        array $annotations
     ): AbstractSchemaBuilder {
-        $annotations = $this->getPropertyAnnotations($propertyReflection);
-
         if ($annotation = $this->findAnnotation($annotations, OA\EnumValue::class)) {
             /** @var OA\EnumValue $annotation */
             if ($annotation->enum) {
@@ -167,12 +163,10 @@ class AnnotationPropertySchemaDecorator implements PropertySchemaDecoratorInterf
         return $builder;
     }
 
-    public function decorateArraySchemaBuilder(
+    private function decorateArraySchemaBuilder(
         ArraySchemaBuilder $builder,
-        ReflectionProperty $propertyReflection
+        array $annotations
     ): AbstractSchemaBuilder {
-        $annotations = $this->getPropertyAnnotations($propertyReflection);
-
         if ($annotation = $this->findAnnotation($annotations, OA\ArrayValue::class)) {
             /** @var OA\ArrayValue $annotation */
             if ($annotation->minItems !== null) {
@@ -189,12 +183,10 @@ class AnnotationPropertySchemaDecorator implements PropertySchemaDecoratorInterf
         return $builder;
     }
 
-    public function decorateHashmapSchemaBuilder(
+    private function decorateHashmapSchemaBuilder(
         HashmapSchemaBuilder $builder,
-        ReflectionProperty $propertyReflection
+        array $annotations
     ): AbstractSchemaBuilder {
-        $annotations = $this->getPropertyAnnotations($propertyReflection);
-
         if ($annotation = $this->findAnnotation($annotations, OA\HashmapValue::class)) {
             /** @var OA\HashmapValue $annotation */
             if ($annotation->requiredProperties !== null) {
@@ -203,35 +195,5 @@ class AnnotationPropertySchemaDecorator implements PropertySchemaDecoratorInterf
         }
 
         return $builder;
-    }
-
-    private function getPropertyAnnotations(ReflectionProperty $propertyReflection): array
-    {
-        return $this->annotationReader->getPropertyAnnotations($propertyReflection);
-    }
-
-    private function findAnnotation(
-        array $annotations,
-        string $annotationClass,
-        bool $exceptionOnAnotherValueInterface = true
-    ): ?object {
-        $found = null;
-        foreach ($annotations as $annotation) {
-            if (\get_class($annotation) === $annotationClass) {
-                // micro-optimalization
-                if (! $exceptionOnAnotherValueInterface) {
-                    return $annotation;
-                }
-
-                $found = $annotation;
-            } elseif (
-                $exceptionOnAnotherValueInterface
-                && \is_subclass_of($annotation, OA\ValueInterface::class)
-            ) {
-                throw new LogicException(\sprintf("Unexpected annotation '%s'", \get_class($annotation)));
-            }
-        }
-
-        return $found;
     }
 }
