+++ src/SchemaBuilder/Decorator/PropertySchemaDecoratorInterface.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ScrumWorks\OpenApiSchema\SchemaBuilder\Decorator;
+
+use ReflectionProperty;
+use ScrumWorks\OpenApiSchema\ValueSchema\Builder\AbstractSchemaBuilder;
+
+interface PropertySchemaDecoratorInterface
+{
+    public function decoratePropertySchemaBuilder(
+        AbstractSchemaBuilder $builder,
+        ReflectionProperty $propertyReflection
+    ): AbstractSchemaBuilder;
+}
