--- src/SchemaBuilder/PropertyDecorator/PropertySchemaDecoratorDefaultTrait.php
@@ -1,94 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace ScrumWorks\OpenApiSchema\SchemaBuilder\PropertyDecorator;
-
-use ReflectionProperty;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\AbstractSchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\ArraySchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\BooleanSchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\EnumSchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\FloatSchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\HashmapSchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\IntegerSchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\MixedSchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\ObjectSchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\StringSchemaBuilder;
-
-trait PropertySchemaDecoratorDefaultTrait
-{
-    /**
-     * General decoration method which run over every ValueScheme.
-     * Useful for solving f.e. descriptions etc.
-     */
-    public function decorateValueSchemaBuilder(
-        AbstractSchemaBuilder $builder,
-        ReflectionProperty $propertyReflection
-    ): AbstractSchemaBuilder {
-        return $builder;
-    }
-
-    public function decorateMixedSchemaBuilder(
-        MixedSchemaBuilder $builder,
-        ReflectionProperty $propertyReflection
-    ): AbstractSchemaBuilder {
-        return $builder;
-    }
-
-    public function decorateIntegerSchemaBuilder(
-        IntegerSchemaBuilder $builder,
-        ReflectionProperty $propertyReflection
-    ): AbstractSchemaBuilder {
-        return $builder;
-    }
-
-    public function decorateFloatSchemaBuilder(
-        FloatSchemaBuilder $builder,
-        ReflectionProperty $propertyReflection
-    ): AbstractSchemaBuilder {
-        return $builder;
-    }
-
-    public function decorateBooleanSchemaBuilder(
-        BooleanSchemaBuilder $builder,
-        ReflectionProperty $propertyReflection
-    ): AbstractSchemaBuilder {
-        return $builder;
-    }
-
-    public function decorateStringSchemaBuilder(
-        StringSchemaBuilder $builder,
-        ReflectionProperty $propertyReflection
-    ): AbstractSchemaBuilder {
-        return $builder;
-    }
-
-    public function decorateEnumSchemaBuilder(
-        EnumSchemaBuilder $builder,
-        ReflectionProperty $propertyReflection
-    ): AbstractSchemaBuilder {
-        return $builder;
-    }
-
-    public function decorateArraySchemaBuilder(
-        ArraySchemaBuilder $builder,
-        ReflectionProperty $propertyReflection
-    ): AbstractSchemaBuilder {
-        return $builder;
-    }
-
-    public function decorateHashmapSchemaBuilder(
-        HashmapSchemaBuilder $builder,
-        ReflectionProperty $propertyReflection
-    ): AbstractSchemaBuilder {
-        return $builder;
-    }
-
-    public function decorateObjectSchemaBuilder(
-        ObjectSchemaBuilder $builder,
-        ReflectionProperty $propertyReflection
-    ): AbstractSchemaBuilder {
-        return $builder;
-    }
-}
