+++ src/SchemaBuilder/SchemaBuilderDecorator.php
@@ -6,17 +6,10 @@ namespace ScrumWorks\OpenApiSchema\SchemaBuilder;
 
 use ReflectionClass;
 use ReflectionProperty;
-use ScrumWorks\OpenApiSchema\Exception\LogicException;
+use ScrumWorks\OpenApiSchema\SchemaBuilder\Decorator\ClassSchemaDecoratorInterface;
+use ScrumWorks\OpenApiSchema\SchemaBuilder\Decorator\PropertySchemaDecoratorInterface;
 use ScrumWorks\OpenApiSchema\ValueSchema\Builder\AbstractSchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\ArraySchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\BooleanSchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\EnumSchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\FloatSchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\HashmapSchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\IntegerSchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\MixedSchemaBuilder;
 use ScrumWorks\OpenApiSchema\ValueSchema\Builder\ObjectSchemaBuilder;
-use ScrumWorks\OpenApiSchema\ValueSchema\Builder\StringSchemaBuilder;
 
 class SchemaBuilderDecorator
 {
@@ -41,33 +34,11 @@ class SchemaBuilderDecorator
     }
 
     public function decorateClassSchemaBuilder(
-        AbstractSchemaBuilder $builder,
+        ObjectSchemaBuilder $builder,
         ReflectionClass $classReflection
     ): AbstractSchemaBuilder {
         foreach ($this->classDecorators as $propertyDecorator) {
-            if ($builder instanceof MixedSchemaBuilder) {
-                $builder = $propertyDecorator->decorateMixedSchemaBuilder($builder, $classReflection);
-            } elseif ($builder instanceof IntegerSchemaBuilder) {
-                $builder = $propertyDecorator->decorateIntegerSchemaBuilder($builder, $classReflection);
-            } elseif ($builder instanceof FloatSchemaBuilder) {
-                $builder = $propertyDecorator->decorateFloatSchemaBuilder($builder, $classReflection);
-            } elseif ($builder instanceof BooleanSchemaBuilder) {
-                $builder = $propertyDecorator->decorateBooleanSchemaBuilder($builder, $classReflection);
-            } elseif ($builder instanceof StringSchemaBuilder) {
-                $builder = $propertyDecorator->decorateStringSchemaBuilder($builder, $classReflection);
-            } elseif ($builder instanceof EnumSchemaBuilder) {
-                $builder = $propertyDecorator->decorateEnumSchemaBuilder($builder, $classReflection);
-            } elseif ($builder instanceof ArraySchemaBuilder) {
-                $builder = $propertyDecorator->decorateArraySchemaBuilder($builder, $classReflection);
-            } elseif ($builder instanceof HashmapSchemaBuilder) {
-                $builder = $propertyDecorator->decorateHashmapSchemaBuilder($builder, $classReflection);
-            } elseif ($builder instanceof ObjectSchemaBuilder) {
-                $builder = $propertyDecorator->decorateObjectSchemaBuilder($builder, $classReflection);
-            } else {
-                throw new LogicException('Unknown schema builder type: ' . \get_class($builder));
-            }
-
-            $builder = $propertyDecorator->decorateValueSchemaBuilder($builder, $classReflection);
+            $builder = $propertyDecorator->decorateObjectSchemaBuilder($builder, $classReflection);
         }
 
         return $builder;
@@ -78,29 +49,7 @@ class SchemaBuilderDecorator
         ReflectionProperty $propertyReflection
     ): AbstractSchemaBuilder {
         foreach ($this->propertyDecorators as $propertyDecorator) {
-            if ($builder instanceof MixedSchemaBuilder) {
-                $builder = $propertyDecorator->decorateMixedSchemaBuilder($builder, $propertyReflection);
-            } elseif ($builder instanceof IntegerSchemaBuilder) {
-                $builder = $propertyDecorator->decorateIntegerSchemaBuilder($builder, $propertyReflection);
-            } elseif ($builder instanceof FloatSchemaBuilder) {
-                $builder = $propertyDecorator->decorateFloatSchemaBuilder($builder, $propertyReflection);
-            } elseif ($builder instanceof BooleanSchemaBuilder) {
-                $builder = $propertyDecorator->decorateBooleanSchemaBuilder($builder, $propertyReflection);
-            } elseif ($builder instanceof StringSchemaBuilder) {
-                $builder = $propertyDecorator->decorateStringSchemaBuilder($builder, $propertyReflection);
-            } elseif ($builder instanceof EnumSchemaBuilder) {
-                $builder = $propertyDecorator->decorateEnumSchemaBuilder($builder, $propertyReflection);
-            } elseif ($builder instanceof ArraySchemaBuilder) {
-                $builder = $propertyDecorator->decorateArraySchemaBuilder($builder, $propertyReflection);
-            } elseif ($builder instanceof HashmapSchemaBuilder) {
-                $builder = $propertyDecorator->decorateHashmapSchemaBuilder($builder, $propertyReflection);
-            } elseif ($builder instanceof ObjectSchemaBuilder) {
-                $builder = $propertyDecorator->decorateObjectSchemaBuilder($builder, $propertyReflection);
-            } else {
-                throw new LogicException('Unknown schema builder type: ' . \get_class($builder));
-            }
-
-            $builder = $propertyDecorator->decorateValueSchemaBuilder($builder, $propertyReflection);
+            $builder = $propertyDecorator->decoratePropertySchemaBuilder($builder, $propertyReflection);
         }
 
         return $builder;
