+++ tests/SchemaParserTest.php
@@ -8,8 +8,8 @@ use Doctrine\Common\Annotations\AnnotationReader;
 use PHPUnit\Framework\TestCase;
 use ScrumWorks\OpenApiSchema\Annotation as OA;
 use ScrumWorks\OpenApiSchema\Exception\LogicException;
-use ScrumWorks\OpenApiSchema\SchemaBuilder\ClassDecorator\AnnotationClassSchemaDecorator;
-use ScrumWorks\OpenApiSchema\SchemaBuilder\PropertyDecorator\AnnotationPropertySchemaDecorator;
+use ScrumWorks\OpenApiSchema\SchemaBuilder\Decorator\ClassDecorator\AnnotationClassSchemaDecorator;
+use ScrumWorks\OpenApiSchema\SchemaBuilder\Decorator\PropertyDecorator\AnnotationPropertySchemaDecorator;
 use ScrumWorks\OpenApiSchema\SchemaBuilder\SchemaBuilderDecorator;
 use ScrumWorks\OpenApiSchema\SchemaBuilder\SchemaBuilderFactory;
 use ScrumWorks\OpenApiSchema\SchemaParser;
